Refactor main; Add save and load buttons
